requirements.txt
python-telegram-bot==20.7
pydub==0.25.1
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
